Provide a generic framework for adding modules.
New modules should go into the scripts directory. These are sourced by
the update script before config files are read, and should provide
functions for the config files to run.

Scripts that are not executable will not be sourced.

This patch adds the add_argument function, which takes three arguments:
 - The command line argument to match against.
 - A bash string to execute with 'eval' upon finding this argument. The
   first argument after the given one will be in $1; make sure to shift
   if you use this.
 - A help string for this argument; this is displayed when the update
   script is invoked with -h.

This patch uses the new framework to split the git support into a
seperate file, and add the '-g' flag to disable it.